// react-module/src/openapi/models/possible-operation-path-to-copy-from.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface PossibleOperationPathToCopyFrom
 */
export interface PossibleOperationPathToCopyFrom {
    /**
     * 
     * @type {number}
     * @memberof PossibleOperationPathToCopyFrom
     */
    completedProductId?: any;
    /**
     * 
     * @type {number}
     * @memberof PossibleOperationPathToCopyFrom
     */
    operationCount?: any;
    /**
     * 
     * @type {AtomicProductExtendedDto}
     * @memberof PossibleOperationPathToCopyFrom
     */
    resultProduct?: any;
}