// react-module/src/openapi/models/tree-node-atomic-operation-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface TreeNodeAtomicOperationExtendedDto
 */
export interface TreeNodeAtomicOperationExtendedDto {
    /**
     * 
     * @type {AtomicOperationExtendedDto}
     * @memberof TreeNodeAtomicOperationExtendedDto
     */
    node?: any;
    /**
     * 
     * @type {Array&lt;TreeNodeAtomicOperationExtendedDto&gt;}
     * @memberof TreeNodeAtomicOperationExtendedDto
     */
    subNodes?: any;
}